Change sntp default update interval to 3 hours.
## Summary
Change sntp default update interval from 1 to 6 hours.

## Impact
This PR aims at reducing the amount of requests sent to the (free as-in-beer and community run) ntp time servers.
The amount of requests for the ntp project should be a a lot less (per esp32) than with the current settings. 

## Rationale
Just from my home network with 5 esp32 controllers there are more than 100 ntp requests per day with the 1 hour update interval. 
The [ntp project](http://www.ntp.org/) is a community run service and open source projects should strive to minimize their impact on this service.