Add espressif/lan867x driver
To be able to support Arduino development on the
[Silicognition ManT1S](https://www.crowdsupply.com/silicognition/mant1s)
board, support for the espressif/lan867x driver first needs to
be added to the Arduino Library.

Support for the ManT1S can then be added later to the ESP32
Arduino Core.

This PR adds generic LAN867x driver support for chips that support
the RMII PHY interface (ESP32 and ESP32P4).